Début session

+++ src/main/java/be/technifutur/labofinal/services/impl/SubclassServiceImpl.java
@@ -1,20 +1,25 @@
 package be.technifutur.labofinal.services.impl;
 
+import be.technifutur.labofinal.exceptions.ElementUsedOnCharacterException;
 import be.technifutur.labofinal.exceptions.ResourceNotFoundException;
 import be.technifutur.labofinal.models.entities.Subclass;
+import be.technifutur.labofinal.repositories.CharacterRepository;
 import be.technifutur.labofinal.repositories.SubclassRepository;
 import be.technifutur.labofinal.services.SubclassService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class SubclassServiceImpl implements SubclassService {
 
     private final SubclassRepository subclassRepository;
+    private final CharacterRepository characterRepository;
 
-    public SubclassServiceImpl(SubclassRepository subclassRepository) {
+    public SubclassServiceImpl(SubclassRepository subclassRepository, CharacterRepository characterRepository) {
         this.subclassRepository = subclassRepository;
+        this.characterRepository = characterRepository;
     }
 
     @Override
@@ -40,6 +45,13 @@ public class SubclassServiceImpl implements SubclassService {
 
     @Override
     public void delete(Long id) {
+        List<Long> characterSubclassId = characterRepository.findAll().stream()
+                .filter(character -> Objects.equals(character.getSubclass().getId(), id))
+                .map(character -> character.getSubclass().getId())
+                .toList();
+        if (!characterSubclassId.isEmpty()) {
+            throw new ElementUsedOnCharacterException(characterSubclassId);
+        }
         subclassRepository.deleteById(id);
     }
 }
